+++ lib/karafka/patches/ruby_kafka.rb
@@ -4,7 +4,7 @@ module Karafka
   module Patches
     module RubyKafka
       # This patch allows us to inject business logic in between fetches and before the consumer
-      # stop, so we can perform shutdown commit or anything else that we need since
+      # stop, so we can perform stop commit or anything else that we need since
       # ruby-kafka fetch loop does not allow that directly
       # We don't wan't to use poll ruby-kafka api as it brings many more problems that we would
       # have to take care of. That way, nothing like that ever happens but we get the control
@@ -12,10 +12,19 @@ module Karafka
       # thread)
       def consumer_loop
         super do
+          controllers = Karafka::Persistence::Controller
+            .all
+            .values
+            .flat_map(&:values)
+            .select { |ctrl| ctrl.respond_to?(:run_callbacks) }
+
           if Karafka::App.stopped?
-            Karafka::Persistence::MessagesConsumer.read.stop
+            controllers.each { |ctrl| ctrl.run_callbacks :before_stop }
+            Karafka::Persistence::Consumer.read.stop
           else
+            controllers.each { |ctrl| ctrl.run_callbacks :before_poll }
             yield
+            controllers.each { |ctrl| ctrl.run_callbacks :after_poll }
           end
         end
       end